+++ lalamo/speculator/common.py
